laft/pyproject.toml
[project]
name = "laft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ "pyusb" ]

[dependency-groups]
dev = [
    "ruff>=0.14.6",
]
